Rename datadog_fmt.rs to json_fmt.rs
The formatter outputs standard flat JSON compatible with Datadog,
Grafana, Elastic, and other platforms — not Datadog-specific.
